# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MAXX-XMD WhatsApp Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/send:
    post:
      operationId: sendMessage
      tags: [bot]
      summary: Send a message via the main bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    SendMessageRequest:
      type: object
      properties:
        number:
          type: string
        message:
          type: string
      required:
        - number
        - message
...